requirements.txt
flask
mysql-connector
gunicorn
flask_restful
pandas
matplotlib
scipy
sklearn
numpy
seaborn
nltk
surprise

